Communication test with LNA board and dewar board done!
The ReceiverControl class it's able to communicate either with
the LNA board and the dewar one. The conversions are right and I need
just to define the methods that perform the remaining requests to the board.
Do so, I need to synchronize the requests, so I could use a mutex on the
low level class (MicroControllerBoard) to avoid
conflicts using the socket at the same time for different methods.
I also need to manage the conversion in a different way than the static one
we are adopting. An idea will be to take a function pointer as a parameter of
the methods that perform the conversion; that function perfoms the conversion.