Added CUSTOM_OPT configuration and related command and tests
I changed the position terminology in order to add this new configuration. 
Now the position is computed as: P = Pis + Pip + Pdp, where Pis is the 
initial static position, Pip the initial parallactic position (computed
using the azimuth and elevation at the beginning of the scan) and Pdp is
the delta parallactic angle between the actual value and Pip.