Added mscu-runserver command
Now to use the MinorServos in test mode we just need to change
the ACS_CDB to the test CDB, and then run the mscu-runserver command.
The unit tests otherwise do not require a live server, because
each test runs it own server by importing the mscu and executing
MSCU.run().