Partial resolution of the noise mark error.
This is a partial fix, not yet tested at the antenna.
The mutex now cover all the calOn() and calOff() code.
I also added the mutex to the updateNoiseMark() method.
I also implemented a workaround to avoid isolated
IO errors: if I got a noise mark error, I do not put 
it a log immediately, but I wait for a second error.
This update will be merged from the trunk to the 
Nuraghe 0.6 TAG.

The next step will affect Nuraghe 0.7. I will enable
the checksum to default. In the MicroControllerBoard
library I will put the checksum control before the 
other checks, and in case of checksum error, I will
raise a new particular exception (MicroControllerBoardIOEx).
When the receiver component catches a MicroControllerBoardIOEx,
it just writes a warning message instead of an error one.

Trunk revision r1485.