Fix #152: send a signal.SIGTERM
I executed the process in the shell and used a process group to send
a signal to all the process in the groups.
I attached a session id to the parent process of the spawned/child
processes, which is a shell in this case. This will make it the group
leader of the processes.  So now, when a signal is sent to the process
group leader, it's transmitted to all of the child processes of this group.