Fix #411, splitted WeatherStation::getData in two different methods. See below. (#412)
* Splitted WeatherStation::getData in two different methods. See below.

The two methods are `getData`, which returns the last read weather parameters immediately, and `updateData`, which performs the request to the devIOs tu update every weather parameter.
The `updateData` method is now called periodically (every 2 seconds) from the `UpdatingThread` (the former `WindCheckerThread`). This thread also checks if the wind is too strong and automatically stows the antenna.

* Fix #411, set `KeepAliveTime=-1` to the weather station component

This will prevent the component to stop updating the weather parameters when it is not referenced by any other component/client.

* Fix #411, updated NotoWeatherStation as well.

The upgrades on Noto code are basically the same as the GenericWeatherStation component.
Also, fixed some typos on common code and a small bug relative to the period of the weather update thread.