Fix #455, splitted the SRT AS GUI into a management one and a view-only one. (#456)
The management one has the buttons to control the functions of the USDs enabled, the view-only one has those buttons disabled and hidden. The view-only one can actually set the profile of the AS or park it like before (is not actually completely view-only, but the USDs cannot be controlled manually).