Fix #477, removed the SecureArea from some AntennaBoss methods (#478)
The said methods are the following:
getObservedHorizontal
getObservedEquatorial
getObservedGalactic
getRawHorizontal
The SecureArea was removed because the TimeTaggedCircularArray objects that these methods use to retrieve the desired coordinates are thread safe objects that do not need further, higher level, syncronization. This will avoid unwanted delays every time one of the above methods is called when the AntennaBoss is sending new coordinates to the mount. This affects all the stations.