Fix #481, implemented the `SRTActiveSurfaceUSDServer::getStatus()` method (#483)
This method returns in a long variable passed by reference, the last known status of the actuator without any call to the actual hardware. This is a measure to prevent occupying the LAN socket in case of tracking. Since during the tracking process the status is updated every time the USD has to move to a different position, I think is safe to assume that the returned status has been read no more than a couple seconds before.