Fix #487, added a `WorkingThreadTime` parameter to SRT AS Boss (#488)
* Fix #487, added a `WorkingThreadTime` parameter to SRT AS Boss

Also, the working thread has been slightly modified in order to take exactly `WorkingThreadTime` to be executed. The standard ACS implementation sets the thread to sleep for the given time, but the total execution time for each iteration becomes `Tex + Tsl` with `Tsl` equal to `WorkingThreadTime`, and `Tex` the total time of execution of a single `runLoop()` method. With this default implementation the total time for each iteration easily exceeds `WorkingThreadTime`. With the new implementation, `Tex` is taken into account so that `Tsl = WorkingThreadTime - Tex`. In this way the total time for each iteration is indeed `WorkingThreadTime` (with fluctuations due to the CPU clock, of course).

* Fix #487, refined thread sleep strategy