Fix #341, fix #507 (#508)
Issue #341: added new nodding configurations for Sardara component. This will be tested on the upcoming 29th of January.
Issue #507: commented the Sardara `setup` command inside component `Init` method. This means that whenever the component starts it no longer sets a default configuration to the Sardara backend. The command with the desired configuration should still be issued by the observer as usual.