fix issue #538: changed the name of the python program in order to run it through a bash script. The script (#544)
turns off the stdout messages in order to provide a clearer output to the user. Improved user messages in case of problems.
slighly changed the default beviour when linking to a component. Default is the calibration tool, if the currently selected
recorder is different a warning is given but the GUI is started all the same.